creating and using alertComposable instead of alertMixin. Moving mappings to it's own file since it's not really a constant that will be used in other components.
